requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
crispy==0.6.3
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.12.0
django-filer==2.0.2
django-js-asset==1.2.2
django-libs==2.0.3
django-mptt==0.12.0
django-phonenumber-field==5.2.0
django-polymorphic==3.0.0
djutils==1.0.13
easy-thumbnails==2.7.1
fabio==0.12.0
greenlet==1.1.1
h5py==3.3.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
oauthlib==3.1.1
phonenumbers==8.12.29
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
silx==0.15.2
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
Unidecode==1.1.2
urllib3==1.26.6
